24 Moved Top Bar to Own File
