+++ .gitignore
@@ -24,3 +24,6 @@ yarn-error.log*
 
 # storybook
 /storybook-static
+
+# local
+/.log
